// libs/enums/property.enum.ts
export enum PropertyType {
  SEDAN = 'SEDAN',
  SUV = 'SUV',
  COUPE = 'COUPE',
  HATCHBACK = 'HATCHBACK',
  CONVERTIBLE = 'CONVERTIBLE',
  VAN = 'VAN',
  TRUCK = 'TRUCK',
  ELECTRIC = 'ELECTRIC',
  WAGON = 'WAGON',
  PICKUP = 'PICKUP',
  MINIVAN = 'MINIVAN',
  CROSSOVER = 'CROSSOVER',
}
// ...
export enum PropertyBrand {
  TOYOTA = 'TOYOTA',
  BMW = 'BMW',
  MERCEDES = 'MERCEDES',
  AUDI = 'AUDI',
  HONDA = 'HONDA',
  FORD = 'FORD',
  CHEVROLET = 'CHEVROLET',
  NISSAN = 'NISSAN',
  HYUNDAI = 'HYUNDAI',
  KIA = 'KIA',
  VOLKSWAGEN = 'VOLKSWAGEN',
  TESLA = 'TESLA',
  LEXUS = 'LEXUS',
  MAZDA = 'MAZDA',
  SUBARU = 'SUBARU',
  JEEP = 'JEEP',
  RAM = 'RAM',
  GMC = 'GMC',
  PORSCHE = 'PORSCHE',
  LAND_ROVER = 'LAND_ROVER',
  VOLVO = 'VOLVO',
  INFINITI = 'INFINITI',
  ACURA = 'ACURA',
  CADILLAC = 'CADILLAC',
  BUICK = 'BUICK',
  CHRYSLER = 'CHRYSLER',
  DODGE = 'DODGE',
  MITSUBISHI = 'MITSUBISHI',
  GENESIS = 'GENESIS',
  ALFA_ROMEO = 'ALFA_ROMEO',
  MINI = 'MINI',
  JAGUAR = 'JAGUAR',
  MASERATI = 'MASERATI',
  BENTLEY = 'BENTLEY',
  ROLLS_ROYCE = 'ROLLS_ROYCE',
  FERRARI = 'FERRARI',
  LAMBORGHINI = 'LAMBORGHINI',
  ASTON_MARTIN = 'ASTON_MARTIN',
  MCLAREN = 'MCLAREN',
  BUGATTI = 'BUGATTI',
  OTHER = 'OTHER',
}
// ...
export enum PropertyLocation {
  MADRID = 'MADRID',
  BARCELONA = 'BARCELONA',
  VALENCIA = 'VALENCIA',
  SEVILLA = 'SEVILLA',
  MALAGA = 'MALAGA',
  BILBAO = 'BILBAO',
  ZARAGOZA = 'ZARAGOZA',
  MURCIA = 'MURCIA',
  ALICANTE = 'ALICANTE',
  GRANADA = 'GRANADA',
  CORDOBA = 'CORDOBA',
  PALMA = 'PALMA',
}